+++ lib/adjust_config.dart
@@ -43,6 +43,7 @@ class AdjustConfig {
   static const String UrlStrategyIndia = 'india';
   static const String UrlStrategyChina = 'china';
   static const String UrlStrategyCn = 'cn';
+  static const String UrlStrategyCnOnly = 'cn-only';
 
   static const String DataResidencyEU = 'data-residency-eu';
   static const String DataResidencyTR = 'data-residency-tr';
@@ -56,6 +57,8 @@ class AdjustConfig {
   static const String AdRevenueSourceUnity = 'unity_sdk';
   static const String AdRevenueSourceHeliumChartboost = 'helium_chartboost_sdk';
   static const String AdRevenueSourcePublisher = 'publisher_sdk';
+  static const String AdRevenueSourceTopOn = 'topon_sdk';
+  static const String AdRevenueSourceAdx = 'adx_sdk';
 
   String _appToken;
   AdjustEnvironment _environment;
@@ -82,6 +85,7 @@ class AdjustConfig {
   bool? coppaCompliantEnabled;
   bool? linkMeEnabled;
   bool? finalAndroidAttributionEnabled;
+  bool? readDeviceInfoOnceEnabled;
   String? sdkPrefix;
   String? userAgent;
   String? defaultTracker;
@@ -89,6 +93,7 @@ class AdjustConfig {
   String? urlStrategy;
   String? processName;
   String? preinstallFilePath;
+  String? fbAppId;
   AdjustLogLevel? logLevel;
   AttributionCallback? attributionCallback;
   SessionSuccessCallback? sessionSuccessCallback;
@@ -219,6 +224,12 @@ class AdjustConfig {
     if (externalDeviceId != null) {
       configMap['externalDeviceId'] = externalDeviceId;
     }
+    if (preinstallFilePath != null) {
+      configMap['preinstallFilePath'] = preinstallFilePath;
+    }
+    if (fbAppId != null) {
+      configMap['fbAppId'] = fbAppId;
+    }
     if (urlStrategy != null) {
       configMap['urlStrategy'] = urlStrategy;
     }
@@ -246,6 +257,9 @@ class AdjustConfig {
     }
     if (finalAndroidAttributionEnabled != null) {
       configMap['finalAndroidAttributionEnabled'] = finalAndroidAttributionEnabled.toString();
+    }
+    if (readDeviceInfoOnceEnabled != null) {
+      configMap['readDeviceInfoOnceEnabled'] = readDeviceInfoOnceEnabled.toString();
     }
     if (linkMeEnabled != null) {
       configMap['linkMeEnabled'] = linkMeEnabled.toString();